split main.go
