+++ util.go
